requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
python-dotenv==0.19.2
djongo==1.3.6 
dnspython==2.2.1 
pymongo==4.3.2 
sqlparse==0.2.4
pycryptodome==3.15.0